version 1.3
Funções separadas
